move bot module
