+++ src/bot/cli.py
@@ -1,7 +1,7 @@
 from dotenv import find_dotenv
 from dotenv import load_dotenv
 
-from .cb import run_bot
+from .bot import run_bot
 
 
 def main():
